More changes for static_ips playbook
We set vm_power_state to poweredoff in main playbook (static_ips.yml) for the dhcp_pxe role to create vms. This variable is then ignored by the task in the static_ips role when we use govc to power on all vms (workers included). Then in finish_install we again used vmware_guest and the vm_power_state var to try and power "on" the workers. Problem is they are already on from govc and by using the var vm_power_state we are setting them to off now which shuts them down.

The fix

static_ips - power on only the bootstrap and master vms
finish_install - reset the vm_power_state var to poweredon
finish_install - power on worker vms